PPTP-1545 : Amend Main Contact Detail
* Extracted parts concerned with contact details amendment into dedicated controller
