+++ app/uk/gov/hmrc/plasticpackagingtax/registration/views/partials/amendment/primaryContactDetails.scala.html
@@ -31,11 +31,11 @@
 @govukSummaryList(
     SummaryList(
         rows = Seq(
-            viewUtils.summaryListRow("primaryContactDetails.check.fullName", primaryContactDetails.name, Some(pptRoutes.AmendRegistrationController.updateContactName())),
-            viewUtils.summaryListRow("primaryContactDetails.check.jobTitle", primaryContactDetails.jobTitle, Some(pptRoutes.AmendRegistrationController.updateJobTitle())),
-            viewUtils.summaryListRow("primaryContactDetails.check.email", primaryContactDetails.email, Some(pptRoutes.AmendRegistrationController.updateEmail())),
-            viewUtils.summaryListRow("primaryContactDetails.check.phoneNumber", primaryContactDetails.phoneNumber, Some(pptRoutes.AmendRegistrationController.updatePhoneNumber())),
-            viewUtils.summaryListRow("primaryContactDetails.check.address", primaryContactDetails.address.map(viewUtils.extractAddress), Some(pptRoutes.AmendRegistrationController.updateAddress()))
+            viewUtils.summaryListRow("primaryContactDetails.check.fullName", primaryContactDetails.name, Some(pptRoutes.AmendContactDetailsController.updateContactName())),
+            viewUtils.summaryListRow("primaryContactDetails.check.jobTitle", primaryContactDetails.jobTitle, Some(pptRoutes.AmendContactDetailsController.updateJobTitle())),
+            viewUtils.summaryListRow("primaryContactDetails.check.email", primaryContactDetails.email, Some(pptRoutes.AmendContactDetailsController.updateEmail())),
+            viewUtils.summaryListRow("primaryContactDetails.check.phoneNumber", primaryContactDetails.phoneNumber, Some(pptRoutes.AmendContactDetailsController.updatePhoneNumber())),
+            viewUtils.summaryListRow("primaryContactDetails.check.address", primaryContactDetails.address.map(viewUtils.extractAddress), Some(pptRoutes.AmendContactDetailsController.updateAddress()))
         ).filterNot(_.value.content == Empty)
     )
 )
